move unit test to test_graph_optimizations.py
