+++ spmd/compiler/distribute.py
@@ -1,41 +1,39 @@
 import logging
 from dataclasses import dataclass
-from enum import auto, Enum
+from enum import Enum, auto
 from functools import partial
-from typing import cast, Dict, List, Sequence, Set, Tuple
-
-import functorch.compile
+from typing import Dict, List, Sequence, Set, Tuple, cast
 
 import torch
 import torch.fx as fx
 import torch.nn as nn
 from functorch.compile import aot_module, make_boxed_func
-from spmd.tensor import (
-    _CURRENT_DECOMPOSITION_TABLE,
-    _Partial,
-    _redistribute_with_local_tensor,
-    DeviceMesh,
-    distribute_tensor,
-    DTensor,
-    operator_dispatch,
-    Placement,
-    propagate_input_sharding,
-    Replicate,
-    Shard,
-)
 from torch.distributed._spmd.comm_tensor import _get_tracer
 from torch.fx.experimental.proxy_tensor import make_fx, proxy_slot
 from torch.utils._pytree import tree_flatten, tree_map, tree_unflatten
+
+from spmd.tensor import (
+    _CURRENT_DECOMPOSITION_TABLE,
+    DeviceMesh,
+    DTensor,
+    Placement,
+    Replicate,
+    Shard,
+    _Partial,
+    _redistribute_with_local_tensor,
+    distribute_tensor,
+    operator_dispatch,
+    propagate_input_sharding,
+)
 
 from .aot_function_patch import patched_aot_function
 from .distributed_graph import DistributedGraph
 from .graph_utils import OP
 from .log_utils import rank0_info
 
-
 # patch aot_function so that we can pass the full (non-sharded) input to capture the graph
 # pyre-fixme
-functorch._src.aot_autograd.aot_function = patched_aot_function
+torch._functorch.aot_autograd.aot_function = patched_aot_function
 
 
 logger: logging.Logger = logging.getLogger(__name__)
