+++ test/spmd/test_tracing.py
@@ -385,165 +385,6 @@ class TraceModuleTest(DTensorTestBase):
                 or p1.grad.allclose(p2.grad)
             )
 
-    @with_comms
-    def test_overlap_pass_grads(self):
-        # test grads with overlap pass vs without overlap
-        # and vs ddp grads
-        model_layer_size = 4
-        gpu_placement = torch.arange(self.world_size)
-
-        class OverlapModel(nn.Module):
-            def __init__(
-                self, layer_count: int = 4, _with_bias: bool = False
-            ) -> None:
-                super().__init__()
-
-                self.seq = nn.Sequential(
-                    *[
-                        nn.Linear(10, 10, bias=_with_bias)
-                        for _ in range(layer_count)
-                    ]
-                )
-
-            def forward(
-                self, x: torch.Tensor
-            ) -> Union[torch.Tensor, Literal[0]]:
-                return sum([self.seq(x)])
-
-        base_model = OverlapModel(layer_count=model_layer_size).to(
-            self.device_type
-        )
-        overlap_model = deepcopy(base_model).to(self.device_type)
-
-        spmd_base = SPMD(
-            base_model,
-            schema=Schema(
-                mesh=DeviceMesh(self.device_type, gpu_placement),
-                placements=[Replicate()],
-            ),
-            optimize_first_iter=False,
-        )
-
-        spmd_overlap = SPMD(
-            overlap_model,
-            schema=Schema(
-                mesh=DeviceMesh(self.device_type, gpu_placement),
-                placements=[Replicate()],
-            ),
-            optimize_first_iter=True,
-        )
-        ddp_model = DDP(deepcopy(overlap_model)).to(self.device_type)
-
-        input = torch.randn(2, 10).to(self.device_type)
-
-        spmd_base(input).sum().backward()
-        spmd_overlap(input).sum().backward()
-        ddp_model(input).sum().backward()
-
-        # compare overlap vs DDP
-        for i, (p1, p2) in enumerate(
-            zip(ddp_model.parameters(), spmd_overlap.parameters())
-        ):
-
-            assert p1.grad.allclose(  # type: ignore
-                p2.grad / self.world_size
-            ), "Mismatch in resulting grads between DDP and SPMD."
-
-            self.assertTrue(
-                p1.grad.allclose(p2.grad / self.world_size)
-                or p1.grad.allclose(p2.grad)
-            )
-
-        # compare overlap vs no overlap
-        for i, (p1, p2) in enumerate(
-            zip(spmd_base.parameters(), spmd_overlap.parameters())
-        ):
-
-            self.assertTrue(p1.grad.allclose(p2.grad))
-
-    @with_comms
-    def test_overlap_pass_called(self):
-        import torch.fx as fx
-
-        @dataclass
-        class RunOverlapProfile:
-            num_calls: int
-
-        @contextmanager
-        def overlap_profiler() -> Generator[RunOverlapProfile, None, None]:
-
-            original_run_overlap = graph_optimization.run_overlap
-            profile: RunOverlapProfile = RunOverlapProfile(num_calls=0)
-
-            # pyre-ignore[53]
-            def patched_run_overlap(
-                gm: fx.GraphModule,
-            ) -> None:
-                original_run_overlap(gm)
-                profile.num_calls += 1
-                print(
-                    f"Entered patched run overlap *************************************************"
-                )
-
-            try:
-                # pyre-ignore[9]
-                run_overlap = patched_run_overlap
-                print(
-                    f" profiler called ************************************************"
-                )
-                print(f"{profile=}")
-                print(f"{run_overlap=}")
-                print(f"{patched_run_overlap=}")
-                yield profile
-            finally:
-                run_overlap = original_run_overlap
-                print(
-                    f"Exiting profiler *********************************************"
-                )
-
-        with overlap_profiler() as profiler:
-            model_layer_size = 2
-            gpu_placement = torch.arange(self.world_size)
-
-            class OverlapModel(nn.Module):
-                def __init__(
-                    self, layer_count: int = 2, _with_bias: bool = False
-                ) -> None:
-                    super().__init__()
-
-                    self.seq = nn.Sequential(
-                        *[
-                            nn.Linear(10, 10, bias=_with_bias)
-                            for _ in range(layer_count)
-                        ]
-                    )
-
-                def forward(
-                    self, x: torch.Tensor
-                ) -> Union[torch.Tensor, Literal[0]]:
-                    return sum([self.seq(x)])
-
-            base_model = OverlapModel(layer_count=model_layer_size).to(
-                self.device_type
-            )
-
-            spmd_overlap = SPMD(
-                base_model,
-                schema=Schema(
-                    mesh=DeviceMesh(self.device_type, gpu_placement),
-                    placements=[Replicate()],
-                ),
-                optimize_first_iter=True,
-            )
-
-            input = torch.randn(2, 10).to(self.device_type)
-
-            spmd_overlap(input).sum().backward()
-
-        self.assertEqual(
-            profiler.num_calls, 1, "Expected a single call to run_overlap."
-        )
-
 
 if __name__ == "__main__":
     run_tests()
